+++ app/commands/adminapi/__init__.py
@@ -1,3 +1,6 @@
-from app.commands.adminapi.command import AdminAPICommand
+from app.commands.adminapi.command import AdminAPICommand, Config
 
-__all__ = ["AdminAPICommand"]
+__all__ = [
+    "AdminAPICommand",
+    "Config",
+]
