+++ configs/dev/dataapi.yaml
@@ -7,3 +7,4 @@ storage:
   port: 6432
   dbname: hyperleda
   user: hyperleda
+  password: fake
