+++ configs/test/dataapi.yaml
@@ -7,4 +7,4 @@ storage:
   port: 5432
   dbname: hyperleda
   user: hyperleda
-  password: password
+  password: fake
